requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
gunicorn==20.1.0